requirements.txt
tensorflow-gpu==2.9.*
pandas
numpy
scipy
pyyaml